+++ subscription/api/blocking/src/main/java/org/occurrent/subscription/api/blocking/SubscriptionModelLifeCycle.java
@@ -1,5 +1,8 @@
 package org.occurrent.subscription.api.blocking;
 
+import java.util.List;
+import java.util.Optional;
+
 /**
  * Defines life-cycle methods for subscription models and subscriptions.
  */