+++ subscription/mongodb/spring/blocking-competing-consumers/src/main/java/org/occurrent/subscription/mongodb/spring/blocking/ListenerLock.java
@@ -1,0 +1,31 @@
+/*
+ * MIT License
+ *
+ * Copyright (c) 2020 Alec Henninger
+ */
+
+package org.occurrent.subscription.mongodb.spring.blocking;
+
+import org.bson.BsonNumber;
+
+import java.util.Objects;
+
+class ListenerLock {
+    private final long version;
+
+    public ListenerLock(BsonNumber version) {
+        Objects.requireNonNull(version, "fencingToken");
+        this.version = version.longValue();
+    }
+
+    public long version() {
+        return version;
+    }
+
+    @Override
+    public String toString() {
+        return "ListenerLock{" +
+                "fencingToken=" + version +
+                '}';
+    }
+}