+++ subscription/mongodb/spring/blocking-competing-consumers/src/main/java/org/occurrent/subscription/mongodb/spring/blocking/LostLockException.java
@@ -1,0 +1,17 @@
+/*
+ * MIT License
+ *
+ * Copyright (c) 2020 Alec Henninger
+ */
+
+package org.occurrent.subscription.mongodb.spring.blocking;
+
+public class LostLockException extends RuntimeException {
+  public LostLockException(String message) {
+    super(message);
+  }
+
+  public LostLockException(String resource, String listenerId) {
+    super(listenerId + " lost lock to " + resource);
+  }
+}