+++ subscription/mongodb/spring/blocking/src/main/java/org/occurrent/subscription/mongodb/spring/blocking/SpringMongoSubscriptionModelConfig.java
@@ -10,7 +10,7 @@ import static java.util.Objects.requireNonNull;
 /**
  * Configuration for the {@code SpringSubscriptionModel}.
  */
-public class SpringSubscriptionModelConfig {
+public class SpringMongoSubscriptionModelConfig {
 
     final String eventCollection;
     final TimeRepresentation timeRepresentation;
@@ -18,17 +18,17 @@ public class SpringSubscriptionModelConfig {
     final boolean restartSubscriptionsOnChangeStreamHistoryLost;
 
     /**
-     * Create a new instance of {@link SpringSubscriptionModelConfig} with the given settings.
+     * Create a new instance of {@link SpringMongoSubscriptionModelConfig} with the given settings.
      * It will by default use a {@link RetryStrategy} for retries, with exponential backoff starting with 100 ms and progressively go up to max 2 seconds wait time between each retry when reading/saving/deleting the subscription position.
      *
      * @param eventCollection    The collection that contains the events
      * @param timeRepresentation How time is represented in the database, must be the same as what's specified for the EventStore that stores the events.
      */
-    public SpringSubscriptionModelConfig(String eventCollection, TimeRepresentation timeRepresentation) {
+    public SpringMongoSubscriptionModelConfig(String eventCollection, TimeRepresentation timeRepresentation) {
         this(eventCollection, timeRepresentation, RetryStrategy.exponentialBackoff(Duration.ofMillis(100), Duration.ofSeconds(2), 2.0f), false);
     }
 
-    private SpringSubscriptionModelConfig(String eventCollection, TimeRepresentation timeRepresentation, RetryStrategy retryStrategy, boolean restartSubscriptionsOnChangeStreamHistoryLost) {
+    private SpringMongoSubscriptionModelConfig(String eventCollection, TimeRepresentation timeRepresentation, RetryStrategy retryStrategy, boolean restartSubscriptionsOnChangeStreamHistoryLost) {
         requireNonNull(eventCollection, "eventCollection cannot be null");
         requireNonNull(timeRepresentation, TimeRepresentation.class.getSimpleName() + " cannot be null");
         requireNonNull(retryStrategy, RetryStrategy.class.getSimpleName() + " cannot be null");
@@ -39,15 +39,15 @@ public class SpringSubscriptionModelConfig {
     }
 
     /**
-     * Create a new SpringSubscriptionModelConfig by using this static method instead of calling the {@link #SpringSubscriptionModelConfig(String, TimeRepresentation)} constructor.
+     * Create a new SpringSubscriptionModelConfig by using this static method instead of calling the {@link #SpringMongoSubscriptionModelConfig(String, TimeRepresentation)} constructor.
      * Behaves the same as calling the constructor so this is just syntactic sugar.
      *
      * @param eventCollection    The collection that contains the events
      * @param timeRepresentation How time is represented in the database, must be the same as what's specified for the EventStore that stores the events.
      * @return A new instance of {@code SpringSubscriptionModelConfig}
      */
-    public static SpringSubscriptionModelConfig withConfig(String eventCollection, TimeRepresentation timeRepresentation) {
-        return new SpringSubscriptionModelConfig(eventCollection, timeRepresentation);
+    public static SpringMongoSubscriptionModelConfig withConfig(String eventCollection, TimeRepresentation timeRepresentation) {
+        return new SpringMongoSubscriptionModelConfig(eventCollection, timeRepresentation);
     }
 
     /**
@@ -63,8 +63,8 @@ public class SpringSubscriptionModelConfig {
      * @param restartSubscriptionsOnChangeStreamHistoryLost Whether or not to automatically restart a subscription, whose change stream history is lost.
      * @return A new instance of {@code SpringSubscriptionModelConfig}
      */
-    public SpringSubscriptionModelConfig restartSubscriptionsOnChangeStreamHistoryLost(boolean restartSubscriptionsOnChangeStreamHistoryLost) {
-        return new SpringSubscriptionModelConfig(eventCollection, timeRepresentation, retryStrategy, restartSubscriptionsOnChangeStreamHistoryLost);
+    public SpringMongoSubscriptionModelConfig restartSubscriptionsOnChangeStreamHistoryLost(boolean restartSubscriptionsOnChangeStreamHistoryLost) {
+        return new SpringMongoSubscriptionModelConfig(eventCollection, timeRepresentation, retryStrategy, restartSubscriptionsOnChangeStreamHistoryLost);
     }
 
     /**
@@ -73,7 +73,7 @@ public class SpringSubscriptionModelConfig {
      * @param retryStrategy A custom retry strategy to use if the {@code action} supplied to the subscription throws an exception
      * @return A new instance of {@code SpringSubscriptionModelConfig}
      */
-    public SpringSubscriptionModelConfig retryStrategy(RetryStrategy retryStrategy) {
-        return new SpringSubscriptionModelConfig(eventCollection, timeRepresentation, retryStrategy, restartSubscriptionsOnChangeStreamHistoryLost);
+    public SpringMongoSubscriptionModelConfig retryStrategy(RetryStrategy retryStrategy) {
+        return new SpringMongoSubscriptionModelConfig(eventCollection, timeRepresentation, retryStrategy, restartSubscriptionsOnChangeStreamHistoryLost);
     }
 }